feat(aim_network): migrate to AIM Network v0.6.x 29-bit protocol
Rename AimNetwork -> aim_network and rewrite the protocol from v0.4.2
(Pkt, 11-bit IDs) to v0.6.x: Msg with 29-bit extended IDs, class-mask
filtering, new aim_job.h timer helper and aim_catalog.h message catalog,
updated ESP32 (TWAI) and STM32 (bxCAN) backend cores.

Rework the canbus examples: fix swapped listener/publisher header
banners, drive the 1 Hz time sync off g_tick1000, and drop dead locals,
an unused aim_job.h include, and a duplicate trailing comment block.

Move the protocol spec to the repo root as the workspace ground-truth
doc (aim_protocol_v0.6.x_draft.md -> aim_protocol_v0.6.0.md) and fix its
title version to match kNetworkVersionString = "0.6.0".

+++ aim_network/src/aim_can_frame.h
@@ -7,9 +7,10 @@ namespace aim {
 
 /**
  * @brief Raw CAN frame format used by all AimNetwork drivers.
+ * All AIM frames use 29-bit extended IDs; cores hard-code IDE.
  */
 struct Frame {
-  uint16_t id;
+  uint32_t id;
   uint8_t dlc;
   uint8_t data[8];
 };
